fix: refactored export, unset and signal handling to please norminette. Also updated cd function, mostly to handle navigation inside deleted directory routes when using relative paths
