+++ second_try/source_files/setup_and_data_handling/env_helper_functions.c
@@ -6,7 +6,7 @@
 /*   By: tvalimak <Tvalimak@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 12:38:55 by akuburas          #+#    #+#             */
-/*   Updated: 2024/05/07 12:26:44 by tvalimak         ###   ########.fr       */
+/*   Updated: 2024/05/09 11:14:06 by tvalimak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,7 +65,6 @@ int	update_shell_level(t_shelldata *data)
 	char		*temp_string;
 
 	temp = data->env_list;
-
 	while (temp != NULL && ft_strcmp(temp->env_var_name, "SHLVL") != 0)
 		temp = temp->next;
 	if (temp == NULL)
