+++ second_try/test_file.txt
@@ -1,0 +1,42 @@
+cd
+rm -fr testfolder
+ls
+mkdir testfolder
+cd testfolder
+ls
+mkdir testfolder1
+cd testfolder1
+mkdir testfolder2
+cd testfolder2
+rm -fr ../../../testfolder
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+cd ..
+pwd
+cd .
+pwd
+ls
